Fix topUpCard method amount should be float instead of int

+++ src/Endpoint/VirtualCards.php
@@ -66,7 +66,7 @@ final class VirtualCards
     /**
      * Amount in cents
      */
-    public function topUpCard(int $amount, string $cardId, $reference): array
+    public function topUpCard(float $amount, string $cardId, $reference): array
     {
         return ResponseMediator::getContent($this->sdk->getHttpClient()->post("$this->baseUri/topup", [], json_encode(compact('amount', 'cardId', 'reference'))));
     }
